Move auth.repository.*UserRepository* -> user.repository.*UserRepository*
