+++ src/test/java/site/hirecruit/hr/domain/auth/service/impl/UserRegistrationRollbackServiceImplTest.kt
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test
 import site.hirecruit.hr.domain.auth.dto.AuthUserInfo
 import site.hirecruit.hr.domain.auth.dto.OAuthAttributes
 import site.hirecruit.hr.domain.user.entity.Role
-import site.hirecruit.hr.domain.auth.repository.UserRepository
+import site.hirecruit.hr.domain.user.repository.UserRepository
 import site.hirecruit.hr.domain.auth.service.TempUserRegistrationService
 import kotlin.random.Random
